Split Hero in two different client and server modules
